Use markdown table instead of html table
Markdown seems overall easier to maintain than html tables.

However there is some downsides :
 - tables with long text values ends up with really long lines. As it's
quite local I think it is tolerable, but AFAIK there is nothing we can
do to get shorter lines in the source file ;
 - to keep the boostrap styling of tables, "{:.table}" is inserted after
the tables. On the Jekyll site, this will trigger the insertion of class
'table' on the html table. This will show as-is in github preview. Not ideal,
but overall worth it I think.