Replace version 1.11.12 by 1.2.0
I assume that there is a typo and that `1.11.12` should have been `1.1.12`. Plus, while there was a `1.1.12-SNAPSHOT`, it was released as `1.2.0`. This is consistent with the [1.2.0 release note](https://github.com/hcoles/pitest/releases/tag/1.2.0) which contains [#215 Automatic selection of targetClasses in Maven plugin](https://github.com/hcoles/pitest/issues/215).

So this commit change references to `1.11.12` to `1.2.0`. I hope I got it right :sweat_smile: